Populateform function  into seperate file
